requirement.txt.txt
# Core Dependencies
numpy==1.26.4
scipy==1.13.1
pandas==1.3.5
matplotlib==3.9.2
seaborn==0.13.2

# Machine Learning & Deep Learning
torch==2.5.1
torchaudio==0.9.1
torchvision==0.10.1+cu111

# Utilities & Performance Optimization
einops==0.8.0
Bottleneck==1.3.7
numexpr==2.8.7
mkl_fft==1.3.10
mkl_random==1.2.7
mkl-service==2.4.0

# Data Processing
openpyxl==3.1.5
pyyaml==6.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2023.3

# Visualization & Plotting
contourpy==1.2.0
cycler==0.11.0
fonttools==4.51.0
kiwisolver==1.4.4
pillow==10.4.0
pyparsing==3.1.2

# Web & Networking
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10

# Jupyter & Interactive Environment
jupyter==1.1.1
notebook==7.2.2
ipython==8.18.1
ipywidgets==8.1.5
jupyterlab==4.2.5
jupyterlab_widgets==3.0.13
jupyterlab_pygments==0.3.0
jupyter_client==8.6.3
jupyter_console==6.6.3
jupyter_core==5.7.2

# Async & Parallel Processing
anyio==4.6.0
async-lru==2.0.4
nest-asyncio==1.6.0
sniffio==1.3.1

# Security & Hashing
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bleach==6.1.0

# Miscellaneous
tqdm==4.66.5
traitlets==5.14.3
packaging==24.1
setuptools==75.1.0
wheel==0.44.0
six==1.16.0
zipp==3.20.2
